plotting stats start

+++ GeRaF/Simulation.cs
@@ -22,8 +22,8 @@ namespace GeRaF
 		// system state
 		public List<Relay> relays;
 		public Dictionary<int, Dictionary<int, double>> distances;
-		public List<Packet> packetsGenerated = new List<Packet>();
 		public List<Packet> packetsFinished = new List<Packet>();
+		public PacketGenerator packetGenerator = new PacketGenerator();
 
 		// simulation state
 		public double clock;
@@ -39,8 +39,10 @@ namespace GeRaF
 			// init sim state
 			clock = 0;
 			eventQueue = new EventQueue();
+
+			var initialCycleDuration = protocolParameters.t_sleep + protocolParameters.t_listen;
 			eventQueue.Add(new StartEvent() {
-				time = 0,
+				time = initialCycleDuration + protocolParameters.t_delta, // must start AFTER first duty cycle events
 				sim = this,
 				previous = null
 			});
@@ -60,7 +62,6 @@ namespace GeRaF
 
 			// init relays as asleep and schedule random awake
 			relays = new List<Relay>();
-			var initialCycleDuration = protocolParameters.t_sleep + protocolParameters.t_listen;
 			for (int i = 0; i < simulationParameters.n_nodes; i++) {
 				var relay = new Relay {
 					id = i,
@@ -79,7 +80,7 @@ namespace GeRaF
 			}
 
 			// move disconnected
-			Console.WriteLine("Placing relays");
+			//Console.WriteLine("Placing relays");
 			var connected = false;
 			int n_slots = (int)Math.Floor(simulationParameters.area_side / simulationParameters.min_distance) - 1;
 			while (!connected) {
@@ -100,72 +101,73 @@ namespace GeRaF
 					relay.Y = Y_slot * simulationParameters.min_distance + simulationParameters.min_distance;
 				}
 
-				Console.WriteLine("Checking connected relays");
+				//Console.WriteLine("Checking connected relays");
 
 				// calculate new neighbours
 				distances = GraphUtils.Distances(relays);
 				GraphUtils.SetNeighbours(relays, distances);
 				connected = GraphUtils.Connected(relays);
 			}
-			Console.WriteLine("Placed relays");
+			//Console.WriteLine("Placed relays");
 
 			// init debug state
-			debugWriter = new StreamWriter(simulationParameters.debug_file);
-			debugWriter.Write("debug_data=`");
-			debugWriter.WriteLine($"{JsonConvert.SerializeObject(simulationParameters)}");
-			debugWriter.WriteLine("#");
-			debugWriter.WriteLine($"{JsonConvert.SerializeObject(protocolParameters)}");
-			debugWriter.WriteLine("#");
-			foreach (var r in relays) {
-				debugWriter.WriteLine($"{r.id};{r.X};{r.Y};{r.range}");
+			if (simulationParameters.debugType != DebugType.Never) {
+				debugWriter = new StreamWriter(simulationParameters.debug_file);
+				debugWriter.Write("debug_data=`");
+				debugWriter.WriteLine($"{JsonConvert.SerializeObject(simulationParameters)}");
+				debugWriter.WriteLine("#");
+				debugWriter.WriteLine($"{JsonConvert.SerializeObject(protocolParameters)}");
+				debugWriter.WriteLine("#");
+				foreach (var r in relays) {
+					debugWriter.WriteLine($"{r.id};{r.X};{r.Y};{r.range}");
+				}
+				debugWriter.WriteLine("#");
+				debugWriter.WriteLine($"{JsonConvert.SerializeObject(distances)}");
+				debugWriter.WriteLine("#");
 			}
-			debugWriter.WriteLine("#");
-			debugWriter.WriteLine($"{JsonConvert.SerializeObject(distances)}");
-			debugWriter.WriteLine("#");
-
 		}
 
-		public void Run() {
-			double lastPrintPercentage = 0;
+		public double progressRate => this.clock / simulationParameters.max_time;
+
+		public SimulationStats Run() {
 			while (eventQueue.isEmpty == false) {
 				var e = eventQueue.Pop();
 				if (e.time < this.clock) {
 					throw new Exception("No time travel plz");
 				}
 				this.clock = e.time;
-				var percentage = this.clock / simulationParameters.max_time;
-				if (percentage > lastPrintPercentage) {
-					Console.WriteLine($"Simulating ... {lastPrintPercentage * 100:F3}%");
-					lastPrintPercentage += 1 / (double)simulationParameters.percentages;
-				}
 				e.Handle();
 				if (simulationParameters.debugType == DebugType.Always) {
 					DebugEvent.DebugNow(this, false, e);
 				}
 			}
 
-			// stats
-			Console.WriteLine($"Number of packets: {packetsFinished.Count}");
-			var tot = (float)packetsFinished.Count;
-			var success = packetsFinished.Count(p => p.result == Result.Success);
-			var max_sensing = packetsFinished.Count(p => p.result == Result.Abort_max_sensing);
-			var max_cycle = packetsFinished.Count(p => p.result == Result.Abort_max_region_cycle);
-			var no_start = packetsFinished.Count(p => p.result == Result.No_start_relays);
-			var max_sink_rts = packetsFinished.Count(p => p.result == Result.Abort_max_sink_rts);
-			var no_ack = packetsFinished.Count(p => p.result == Result.Abort_no_ack);
-			Console.WriteLine($"Number of packets success: {success}, {success / tot * 100:F3}%");
-			Console.WriteLine($"Number of packets no_start_relays: {no_start}, {no_start / tot * 100:F3}%");
-			Console.WriteLine($"Number of packets max_cycle: {max_cycle}, {max_cycle / tot * 100:F3}%");
-			Console.WriteLine($"Number of packets max_sensing: {max_sensing}, {max_sensing / tot * 100:F3}%");
-			Console.WriteLine($"Number of packets max_sink_rts: {max_sink_rts}, {max_sink_rts / tot * 100:F3}%");
-			Console.WriteLine($"Number of packets no_ack: {no_ack}, {no_ack / tot * 100:F3}%");
-
-			debugWriter.WriteLine("#");
-			foreach (var p in packetsGenerated) {
-				debugWriter.WriteLine($"{JsonConvert.SerializeObject(p)}");
+			if (simulationParameters.debugType != DebugType.Never) {
+				debugWriter.Write("`");
+				debugWriter.Close();
 			}
-			debugWriter.Write("`");
-			debugWriter.Close();
+
+			// stats
+			var simStats = new SimulationStats() {
+				finishedPackets = String.Join(";", packetsFinished.Select(p => p.ToString())),
+				relayInfos = String.Join(";", relays.Select(r => new RelayInfo() {
+					id = r.id,
+					X = r.X,
+					Y = r.Y,
+					asleepTime = 0,
+					awakeTime = 0
+				}.ToString()))
+			};
+
+			return simStats;
+
+			//var tot = (float)packetsFinished.Count;
+			//var success = packetsFinished.Count(p => p.result == Result.Success);
+			//var max_sensing = packetsFinished.Count(p => p.result == Result.Abort_max_sensing);
+			//var max_cycle = packetsFinished.Count(p => p.result == Result.Abort_max_region_cycle);
+			//var no_start = packetsFinished.Count(p => p.result == Result.No_start_relays);
+			//var max_sink_rts = packetsFinished.Count(p => p.result == Result.Abort_max_sink_rts);
+			//var no_ack = packetsFinished.Count(p => p.result == Result.Abort_no_ack);
 		}
 	}
 }
